refactor(RuntimeUpdates): decouple runtime from being constructed from a class method
This will allow for programmatic execution of arbitrary functions without having to rely on scripts.
